Move heap out of stdlib
The heap is not used by all targets.
