+++ stdlib/Makefile
@@ -5,7 +5,7 @@ bpf/%.bc wasm/%.bc: %.c
 	$(CC) -c $(CFLAGS) $< -o $@
 
 SOLANA=$(addprefix bpf/,solana.bc bigint.bc format.bc)
-WASM=$(addprefix wasm/,ripemd160.bc sha3.bc stdlib.bc substrate.bc bigint.bc format.bc)
+WASM=$(addprefix wasm/,ripemd160.bc sha3.bc stdlib.bc substrate.bc bigint.bc format.bc wasmheap.bc)
 
 all: $(SOLANA) $(WASM)
