+++ stdlib/stdlib.h
@@ -11,4 +11,4 @@ struct vector
 void *__malloc(size_t size);
 void __memset(void *dest, uint8_t val, size_t length);
 void __memcpy(void *dest, const void *src, size_t length);
-
+void __memcpy8(void *_dest, void *_src, size_t length);
